requirements.txt
django>=1.10
djangorestframework-jwt==1.11.0
djangorestframework==3.6.4
pyjwt==1.5.3
pytz==2017.2
git+https://github.com/chessbr/rest-jwt-permission@master#egg=rest-jwt-permission
